Added in prototype of echolocation
created a simple version of the echolocation feature where upon clicking a button an indicator appears that follows the mouse and if the mouse button is clicked within a certain area then it shoots out a sprite mask the reveals the sprites of the level. Temporarily commented some yet to be implemented level set up features for testing.